Move CorruptFile() utilityy function to testutil.cc
